requirements.txt
aiogram==2.16
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.3.4
async-timeout==4.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.7
frozenlist==1.2.0
h11==0.12.0
httpcore==0.14.2
httpx==0.21.1
idna==3.3
multidict==5.2.0
python-environ==0.4.54
pytz==2021.3
rfc3986==1.5.0
schedule==1.1.0
sniffio==1.2.0
typing_extensions==4.0.0
yarl==1.7.2
